fix: use data binding correctly for signup section
